pyproject.toml
[tool.poetry]
name = "example"
version = "0.1.0"
description = ""
authors = ["Eugene Morozov <jmv@emorozov.net>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
django-environ = "^0.8.1"
psycopg2 = {extras = ["binary"], version = "^2.9.3"}
djangorestframework = "^3.13.1"
django-extensions = "^3.1.5"
gunicorn = "^20.1.0"
uvicorn = "^0.17.6"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
flake8 = "^4.0.1"
ipython = "^8.3.0"
mypy = "^0.942"
pytest-django = "^4.5.2"
pytest-cov = "^3.0.0"
django-stubs = "^1.10.1"
pre-commit = "^2.19.0"
yesqa = "^1.3.0"
djangorestframework-stubs = "^1.5.0"
autoflake = "^1.4"
isort = "^5.10.1"
httpretty = "^1.1.4"
pytest-socket = "^0.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
